+++ controllers/engines/router.js
@@ -11,30 +11,19 @@ const { _U_ } = require('../utilities/');
 
 // helpers
 /**
- * wrap asyncfunctions with an error catcher                                                                            // note: `asyncWrapper` may be generalized
- * @param {(array|function)} fn             - fn may be wrapped
- * @return {array}                          - task is triggered only when keyword 'async' is met
+ * wrap asyncfunctions with an error catcher
+ * @param {array|function} target           - fn may be wrapped
+ * @return {array|function}                 - task is triggered only when keyword 'async' is found
  */
-const asyncWrapper = (fn) => {
-  const wrapAsync = (fn) => (req, res, next) => fn(req, res, next).catch(next);
-  if (!_U_.object.checkNativeBrand(fn, 'Array')) fn = [fn];
-  return fn.map(fn => _U_.object.checkNativeBrand(fn, 'AsyncFunction') ? wrapAsync(fn) : fn);
+const asyncWrapper = (target) => {
+  const wrapper = (fn) => (req, res, next) => fn(req, res, next).catch(next);
+  const evaluator = (fn) => _U_.object.checkNativeBrand(fn, 'AsyncFunction') ? wrapper(fn) : fn;
+  const type = _U_.object.checkNativeBrand(target);
+  if (type === 'Array') return target.map(fn => evaluator(fn));
+  if (type.toLowerCase().includes('function')) return evaluator(target);
+  throw new TypeError(`Argument is neither an Array nor an AsyncFunction but a ${type}.`);
 };
 
-/**
- * normalize into a method array in the anti-alphabetical order
- * @param {(object|function)} controller    - controller may be extracted
- * @param {(string|regex)} [alias]          - alias to be watched
- * @param {(string|array)} [method]         - default: keys of controller{object} || `get`
- * @return {array}                          - ordering of the result is important ('alias' must be met after 'get')
- */
-const stackHttpMethods = ({ controller, alias, method }) => {
-  const methodList = method
-    ? _U_.object.checkNativeBrand(method, 'String') ? [method] : method
-    : _U_.object.checkNativeBrand(controller, 'Object') ? Object.keys(controller) : ['get'];
-  if (alias && !methodList.includes('alias')) methodList.push('alias');
-  return methodList.sort();
-};
 
 /**
  * stack a queue from setting
@@ -45,10 +34,9 @@ const stackHttpMethods = ({ controller, alias, method }) => {
  * @param {boolean} [crawler]               - accept crawlers? (load 1 fn)                      default: true
  * @param {string} [title]                  - title tag name
  * @param {object} [titleOption]            - title tagging options (see _M_.setTitleTag())
- * @param {string} [method]                 - method to be watched
- * @return {Array}                          - ordering of the result is important
+ * @return {array}                          - ordering of the result is important
  */
-const loadRoutePlugins = ({ query, cache, authorized, authenticated, crawler, title, titleOption } = {}, method) => {
+const getPreprocessor = ({ query, cache, authorized, authenticated, crawler, title, titleOption } = {}) => {
   const queue = [];
   if (query !== 'sensitive') queue.push(_M_.caseInsensitiveQuery);
   if (cache === false) queue.push(_M_.doNotCached);
@@ -59,27 +47,18 @@ const loadRoutePlugins = ({ query, cache, authorized, authenticated, crawler, ti
   return queue;
 };
 
-/**
- * stack a queue from controllers with normalization
- * @param {(array|function)} controller     - controller to be stacked
- * @param {string} [method]                 - method to be used to extract property of the controller
- * @return {array}                          - if no matched method, the controller would just be normalized to an array
- */
-const loadMainControls = (controller, method) => {
-  const agent = _U_.object.proxyfiedForCaseInsensitiveAccess(controller);
-  const worker = agent[method] ? agent[method] : agent;
-  return _U_.object.checkNativeBrand(worker, 'Array') ? worker : [worker];
-};
 
 /**
- * stack a queue with a given template when HTTP method is 'get'
- * @param {string|null} [template]          - template file path
- * @param {string|null} [handler]           - template handler name
- * @param {string} method                   - method to be watched
- * @return {array}                          - task is triggered only when the method is 'alias' or 'get'
+ *
+ * @param {array|function} protagonist      - the main controlling logic
+ * @param {object} hooker                   - pre/post hooked middleware on the device
+ * @param {object} setting                  - router settings
+ * @return {array}
  */
-const loadViewRenderer = ({ template, handler } = {}, method) => {
-  return ['get', 'alias'].includes(method.toLowerCase()) && template ? [templateHandler(template, handler)] : [];
+const getMiddlewareChain = (protagonist, hooker, setting) => {
+  const preprocessor = getPreprocessor(setting);
+  const viewHandler = templateHandler(setting.template, setting.handler);
+  return asyncWrapper([].concat(preprocessor, hooker.pre, protagonist, hooker.post, viewHandler));
 };
 
 
@@ -91,18 +70,18 @@ const loadViewRenderer = ({ template, handler } = {}, method) => {
  * @param {object} [option]                 - (see express.Router() API)
  */
 class Device {
-  constructor(rules, option) {                                                                                          // todo: [private] make these private once JS supports
-    this._router = new Router(option);
-    this._queue = { pre: [], post: [] };
-    this._rules = rules;
-    this._handler = '';
+  constructor(rules, option) {                                                                                         
+    this.router = new Router(option);
+    this.queue = { pre: [], post: [] };
+    this.rules = rules;
+    this.handler = null;
   }
 
   get setting() {
     return new Proxy(this, {
       get: (target, name) => arg => name === 'handler'
-        ? target._handler = arg
-        : target.hook('pre', loadRoutePlugins({ [name]: arg })),
+        ? target.handler = arg
+        : target.hook('pre', getPreprocessor({ [name]: arg })),
     });
   }
 
@@ -111,30 +90,30 @@ class Device {
   }
 
   hook (position, fn) {
-    if(!Array.isArray(fn)) fn = [fn];
-    this._queue[position].push(...fn);
+    this.queue[position].push(...(_U_.object.checkNativeBrand(fn, 'Array') ? fn : [fn]));
     return this;
   }
 
   use(fn) {
-    this._router.use(fn);
+    this.router.use(fn);
     return this;
   }
 
   run() {
-    this._rules.forEach(({ route, alias, controller, method, setting }) => {
-      if (this._handler) setting = { handler: this._handler, ...setting };
-      stackHttpMethods({ alias, controller, method })
-        .forEach(method => this._router[(method === 'alias') ? 'get' : method.toLowerCase()](...asyncWrapper([
-          (method === 'alias') ? alias : route,                                                                         // router path
-          ...loadRoutePlugins(setting   , method),                                                                      // middleware plugins
-          ...this._queue.pre,                                                                                           // universal pre-middleware
-          ...loadMainControls(controller, method),                                                                      // router controller
-          ...loadViewRenderer(setting   , method),                                                                      // template handler
-          ...this._queue.post                                                                                           // universal post-middleware
-        ])));
+    // router registrations
+    this.rules.forEach(({ route, alias, controller, setting }) => {
+      const controlKeys = Object.keys(controller).sort();
+      const settingWrap = { handler: this.handler, ...setting };
+
+      // method registrations
+      controlKeys.forEach(key => {
+        if (key === 'alias' && !alias) throw new ReferenceError('Parameter "alias" have to be provided.');
+        const path = key === 'alias' ? alias : route;
+        const method = key === 'alias' ? 'get' : key.toLowerCase();
+        this.router[method](path, getMiddlewareChain(controller[key], this.queue, settingWrap));
+      });
     });
-    return this._router;
+    return this.router;
   }
 }
 
@@ -144,9 +123,7 @@ class Device {
 module.exports = { Device, _test: {
     Device,
     asyncWrapper,
-    stackHttpMethods,
-    loadRoutePlugins,
-    loadMainControls,
-    loadViewRenderer,
+    getPreprocessor,
+    getMiddlewareChain,
   },
 };